Add event slider to home page
